# docs/swagger.yaml
basePath: /api
definitions:
...
  BanResponse:
    properties:
      id:
        type: string
      image:
        type: string
      username:
        type: string
    type: object
...
  MemberRequest:
    properties:
      memberId:
        type: string
    type: object
  MemberSettings:
    properties:
      color:
        type: string
      nickname:
        type: string
    type: object
...
  model.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/model.HttpError'
    type: object
...
  model.HttpError:
    properties:
      message:
        type: string
      type:
        type: string
    type: object
...
  model.Success:
    properties:
      success:
        type: boolean
    type: object
...
host: localhost:4000
info:
  contact:
    email: khoingoc456@gmail.com
    name: KhoiPN
  description: MurmurChatapp REST API Specs. This service uses sessions for authentication
  title: MurmurChatapp API
  version: "1.0"
paths:
...
  /guilds/{guildId}/bans:
    delete:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      - description: Member ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Success'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Unban Member
      tags:
      - Members
    get:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/BanResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get Guild Ban list
      tags:
      - Members
    post:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      - description: Member ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Success'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Ban Member
      tags:
      - Members
...
  /guilds/{guildId}/kick:
    post:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      - description: Member ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Success'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Kick Member
      tags:
      - Members
  /guilds/{guildId}/member:
    get:
      parameters:
      - description: Guild ID
        in: path
        name: guildId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MemberSettings'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get Member Settings
      tags:
      - Members
...
swagger: "2.0"